backend/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = "General description"
authors = ["Mike Odnis <mikeodnis3242004@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
flask-restx = "^1.3.0"
cloudant = "^2.15.0"
urllib3 = "1.26.19" # DO NOT upgrade this!
retry2 = "^0.9.5"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
honcho = "^1.1.0"
pylint = "^3.2.2"
flake8 = "^7.0.0"
black = "^24.4.2"
pytest = "^8.2.1"
pytest-pspec = "^0.0.4"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.0"
coverage = "^7.6.0"
httpie = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

########################################
# Tool configurations
########################################

[tool.flake8]
# ignore = ['E231', 'E241']
# Ignore module imports for routes
per-file-ignores = [
    '*/__init__.py:F401 E402',
]
max-line-length = 127
count = true

[tool.black]
line-length = 127

[tool.pylint.'MESSAGES CONTROL']
disable = "no-member,protected-access,global-statement"

[tool.pylint.FORMAT]
max-line-length = 127

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--pspec --cov=service --cov-fail-under=95"
testpaths = ["tests"]

[tool.coverage.run]
source = ["service"]
omit = [
    "venv/*",
    ".venv/*"
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "pragma: no branch",
    "pass",
    "subprocess.CalledProcessError",
    "sys.exit",
    "if __name__ == .__main__.:"
]
ignore_errors = true

[tool.coverage.xml]
output="./coverage.xml"

[tool.coverage.html]
title = "Test Coverage Report"
directory = "coverage_html_report"
